FLIZ-452 feat: /notification/disconnect 페이지 네트워크 최적화
